\name{add_class}
\alias{add_class}
\title{add a class attribute}
\usage{
  add_class(x, new.class)
}
\arguments{
  \item{x}{an object to be given a new class}

  \item{new.class}{a character string}
}
\description{
  add_class adds the specified class to the beginning of an
  object's class attribute vector. The object retains all
  previous class memberships.
}
\keyword{internal}

