\name{gid}
\alias{gid}
\title{Create grouping variable}
\usage{
  gid(ens)
}
\arguments{
  \item{ens}{an object of class mg_ensemble}
}
\description{
  gid uses the information in the groups attribute of an
  mg_enemble object to create a grouping variable. This
  variable should be saved as a column in the dataset on
  which group-wise modelling will be performed.
}
\keyword{internal}

