\name{is.grouped}
\alias{is.grouped}
\title{Is x a modelglyphs object?}
\usage{
  is.grouped(x)
}
\arguments{
  \item{x}{An object to be tested for membership in a
  modelglyphs class}
}
\description{
  is.mg tests whether an object inherits from a modelglyphs
  class. Such objects should all contain a data set with a
  gid variable and have the following attributes:
  model_info, key, x_major, y_major.
}

