\name{is.models}
\alias{is.models}
\title{Is an object a models object from the modelglyphs package?}
\usage{
  is.models(x)
}
\arguments{
  \item{x}{An object to be tested}
}
\value{
  logical TRUE if x inherits from the 'models' class, FALSE
  otherwise
}
\description{
  Is an object a models object from the modelglyphs
  package?
}

