\name{mg_call}
\alias{mg_call}
\title{return the model call of an modelglyphs object}
\usage{
  mg_call(mg)
}
\arguments{
  \item{mg}{A modelglyphs object. The class of a
  modelglyphs object will begin with "mg_"}
}
\description{
  mg_call returns the model call of a modelglyphs object as
  a character string.
}

