\name{switch_class}
\alias{switch_class}
\title{switch the first class attribute}
\usage{
  switch_class(x, new.class)
}
\arguments{
  \item{x}{an object to be given a new class}

  \item{new.class}{a character string}
}
\description{
  switch_class replaces the first element of an object's
  class vector with the specified class. The object retains
  all other class memberships.
}
\keyword{internal}

