\name{mg}
\alias{mg}
\title{Add the mg.info attribute of one object to another}
\usage{
  mg(out, inp, class = NULL)
}
\arguments{
  \item{out}{An object to add an mg.info attribute to}

  \item{inp}{An object with an mg.info attribute to copy to
  out}

  \item{class}{A character string to add to the start of
  the new object's class vector.}
}
\description{
  Add the mg.info attribute of one object to another
}
\keyword{internal}

